refactor: Reservation 도메인 로직 패키지 구조 수정
